tlsconfig: move Clone into proper package
