+++ pkg/tlsconfig/tlsconfig_clone_go16.go
@@ -1,12 +1,12 @@
 // +build go1.6,!go1.7
 
-package transport
+package tlsconfig
 
 import "crypto/tls"
 
-// TLSConfigClone returns a clone of tls.Config. This function is provided for
+// Clone returns a clone of tls.Config. This function is provided for
 // compatibility for go1.6 that doesn't include this method in stdlib.
-func TLSConfigClone(c *tls.Config) *tls.Config {
+func Clone(c *tls.Config) *tls.Config {
 	return &tls.Config{
 		Rand:                     c.Rand,
 		Time:                     c.Time,
